+++ src/layout/index.ts
@@ -1,2 +1,3 @@
 export { default as FormRender } from './form-render.vue';
 export { default as ComponentRender } from './component-render.vue';
+export * from './types';