+++ src/modal/index.ts
@@ -1,4 +1,5 @@
 export { default as ModalView } from './df-api.component.vue';
 export { default as DfModal } from './df-modal.component.vue';
+export type { DfModalProps, DfModalSlots } from './df-modal.types';
 export { default as modal, type CloseablePromise, type FormActions, type ModalOptions } from './api';
 export { default as DialogSize, defaultDialogSize } from './dialog-size';